refactor: merge mesh-editor and static-mesh into aether-mesh-viewer-component
Retire aether-mesh-editor-component and aether-static-mesh-component. The
new aether-mesh-viewer-component takes one mail kind, aether.mesh.load
{ namespace, path }, and dispatches on file extension: .dsl runs through
aether-mesh's parser+mesher with polygon-edge wireframes; .obj runs the
fan-triangulation parser with default soft-blue color and no wireframe.

The old editor's inline aether.mesh.set_text path is gone — the agent
loop becomes write-file via aether.io.write, then aether.mesh.load
against the same path. Trades one mail per iteration for one mail kind
total instead of three (set_text, set_path, static_mesh.load).

Smoke-tested live: DSL cube renders with slate wireframe; subsequent OBJ
tetrahedron load atomically replaces the cache.
