+++ Cargo.toml
@@ -10,8 +10,7 @@ members = [
     "crates/aether-component",
     "crates/aether-mesh",
     "crates/aether-camera-component",
-    "crates/aether-mesh-editor-component",
-    "crates/aether-static-mesh-component",
+    "crates/aether-mesh-viewer-component",
     "crates/aether-hub-protocol",
     "crates/aether-mail",
     "crates/aether-mail-derive",
