+++ crates/aether-mesh/src/lib.rs
@@ -2,8 +2,8 @@
 //! primitive-composition format defined by ADR-0026 and ADR-0051.
 //!
 //! Library-only (per ADR-0053): produces triangles, doesn't render. The
-//! `aether-mesh-editor-component` consumes this crate to mesh DSL text
-//! sent over mail; the `dsl_to_obj` example converts a `.dsl` file to
+//! `aether-mesh-viewer-component` consumes this crate to mesh DSL text
+//! loaded from disk; the `dsl_to_obj` example converts a `.dsl` file to
 //! Wavefront OBJ for inspection in any external viewer.
 //!
 //! Boolean composition (`union` / `intersection` / `difference`) was
