+++ crates/aether-mesh/src/polygon.rs
@@ -3,7 +3,7 @@
 //! `Polygon` is the canonical mesh form for the engine — one `Polygon`
 //! per logical face, with optional hole loops for pierced regions.
 //! Triangulation moves to display time via [`tessellate_polygon`], so
-//! consumers (notably `aether-mesh-editor-component`) hand polygons to
+//! consumers (notably `aether-mesh-viewer-component`) hand polygons to
 //! the GPU upload step rather than triangles.
 //!
 //! [`mesh_polygons`] is the polygon-domain analogue of [`crate::mesh`]:
@@ -26,7 +26,7 @@
 //! Vertices are stored as [`Point3`] (16:16 fixed-point integers) end-
 //! to-end through the mesh pipeline — same type the BSP CSG core and
 //! cleanup passes already use. The conversion to `f32` happens at the
-//! GPU upload boundary inside `aether-mesh-editor-component`, not
+//! GPU upload boundary inside `aether-mesh-viewer-component`, not
 //! here. Keeping the polygon-domain integer-typed eliminates the f32
 //! noise that previously caused `is_convex` and CDT to disagree on
 //! near-collinear vertices (issue 335). The only `f32` field is
@@ -318,7 +318,7 @@ fn projection_axes(plane: &plane::Plane3) -> (Axis, Axis) {
 /// Tessellate a polygon (with optional holes) into triangles for GPU
 /// upload. Output triangles are wound CCW around `plane_normal`. This
 /// is the display-time tessellation step ADR-0057 moves out of the
-/// cleanup pipeline — consumers like `aether-mesh-editor-component`
+/// cleanup pipeline — consumers like `aether-mesh-viewer-component`
 /// call it once per polygon when assembling render-ready geometry.
 ///
 /// Output triangles carry [`Point3`] vertices; convert via
